+++ app/abstract-exchange-subscription-manager.js
@@ -3,6 +3,7 @@ const _ = require('lodash');
 const EventEmitter = require('events');
 const debug = require('debug')('CEG:ExchangeSubscriptionManager');
 const logger = require('winston');
+const Errors = require('./errors');
 const AbstractExchangeClass = require('./abstract-exchange');
 
 /**
@@ -16,8 +17,14 @@ const AbstractExchangeClass = require('./abstract-exchange');
  * - trades (new trades for a single pair)
  */
 
-// how often in ms should we run ticker loop
-const TICKER_LOOP_DEFAULT_PERIOD = 30 * 1000;
+// how often in ms should we retrieve tickers
+const TICKERS_LOOP_DEFAULT_PERIOD = 30 * 1000;
+
+// how often in ms should we retrieve order book
+const ORDER_BOOK_LOOP_DEFAULT_PERIOD = 30 * 1000;
+
+// how often in ms should we retrieve trades
+const TRADES_LOOP_DEFAULT_PERIOD = 30 * 1000;
 
 class AbstractExchangeSubscriptionManager extends EventEmitter
 {
@@ -28,7 +35,6 @@ class AbstractExchangeSubscriptionManager extends EventEmitter
  * @param {object} exchange Exchange instance
  * @param {boolean} opt.marketsSubscription indicates exchange support market subscription (ie: orderbook & trades at the same time) (optional, default = true)
  * @param {boolean} opt.globalTickersSubscription indicates exchange support a single subscription for all tickers (optional, default = true)
- * @param {boolean} opt.tickerLoop indicates whether or not ticker loop should be started automatically (optional, default = false)
  */
 constructor(exchange, options)
 {
@@ -39,12 +45,33 @@ constructor(exchange, options)
     super();
     this._globalTickersSubscription = true;
     this._marketsSubscription = true;
-    // whether or not ticker stream should be providing by doing periodic REST requests
-    this._tickerLoop = {
-        enabled:false,
-        list:{},
-        period:TICKER_LOOP_DEFAULT_PERIOD
+    // used to emulate ws by doing periodic REST requests
+    this._emulatedWs = {
+        wsTickers:{
+            enabled:false,
+            list:{},
+            period:TICKERS_LOOP_DEFAULT_PERIOD
+        },
+        wsOrderBooks:{
+            enabled:false,
+            list:{},
+            period:ORDER_BOOK_LOOP_DEFAULT_PERIOD
+        },
+        wsTrades:{
+            enabled:false,
+            list:{},
+            period:TRADES_LOOP_DEFAULT_PERIOD
+        }
     }
+    // check if we need ws emulation
+    let features = exchange.getFeatures();
+    _.forEach(['wsTickers','wsOrderBooks','wsTrades'], (type) => {
+        if (features[type].enabled && features[type].emulated)
+        {
+            this._emulatedWs[type].enabled = true;
+            this._emulatedWs[type].period = features[type].period * 1000;
+        }
+    });
     if (undefined !== options)
     {
         if (false === options.marketsSubscription)
@@ -54,21 +81,6 @@ constructor(exchange, options)
         if (false === options.globalTickersSubscription)
         {
             this._globalTickersSubscription = false;
-        }
-        if (undefined !== options.tickerLoop)
-        {
-            if (true === options.tickerLoop.enabled)
-            {
-                this._tickerLoop.enabled = true;
-            }
-            if (undefined !== options.tickerLoop.period)
-            {
-                let period = parseInt(options.tickerLoop.period);
-                if (!isNaN(period) && period > 0)
-                {
-                    this._tickerLoop.period = period * 1000;
-                }
-            }
         }
     }
     this._exchangeInstance = exchange;
@@ -286,13 +298,13 @@ updateTickersSubscriptions(sessionId, subscribe, unsubscribe, connect)
         }
         this._subscriptions.tickers.timestamp = timestamp;
         this._subscriptions.tickers.count = Object.keys(this._subscriptions.tickers.pairs).length;
-        if (!this._tickerLoop.enabled)
+        if (!this._emulatedWs.wsTickers.enabled)
         {
             this._processChanges(changes, {connect:connect});
         }
         else
         {
-            this._processTickerLoop(changes, {connect:connect});
+            this._processTickersLoops(changes, {connect:connect});
         }
     }
 }
@@ -428,7 +440,14 @@ updateOrderBooksSubscriptions(sessionId, subscribe, unsubscribe, resync, connect
         }
         this._subscriptions.orderBooks.count = Object.keys(this._subscriptions.orderBooks.pairs).length;
         this._subscriptions.markets.count = Object.keys(this._subscriptions.markets.pairs).length;
-        this._processChanges(changes, {connect:connect});
+        if (!this._emulatedWs.wsOrderBooks.enabled)
+        {
+            this._processChanges(changes, {connect:connect});
+        }
+        else
+        {
+            this._processOrderBooksLoops(changes, {connect:connect});
+        }
     }
 }
 
@@ -543,7 +562,14 @@ updateTradesSubscriptions(sessionId, subscribe, unsubscribe, connect)
         this._subscriptions.trades.timestamp = timestamp;
         this._subscriptions.trades.count = Object.keys(this._subscriptions.trades.pairs).length;
         this._subscriptions.markets.count = Object.keys(this._subscriptions.markets.pairs).length;
-        this._processChanges(changes, {connect:connect});
+        if (!this._emulatedWs.wsOrderBooks.enabled)
+        {
+            this._processChanges(changes, {connect:connect});
+        }
+        else
+        {
+            this._processTradesLoops(changes, {connect:connect});
+        }
     }
 }
 
@@ -786,8 +812,8 @@ _processSubscriptions(streamClientDescriptor)
         switch (entity)
         {
             case 'tickers':
-                // ignore as it is already managed by tickerLoop
-                if (this._tickerLoop.enabled)
+                // ignore as it is already managed by tickers loop
+                if (this._emulatedWs.wsTickers.enabled)
                 {
                     return;
                 }
@@ -817,7 +843,7 @@ _processSubscriptions(streamClientDescriptor)
             }
         });
     });
-    if (this._globalTickersSubscription && !this._tickerLoop.enabled)
+    if (this._globalTickersSubscription && !this._emulatedWs.wsTickers.enabled)
     {
         if (this._subscriptions.tickers.subscribed)
         {
@@ -831,7 +857,7 @@ _processSubscriptions(streamClientDescriptor)
     this._processChanges(changes, {connect:true, client:streamClientDescriptor});
 }
 
-//-- ticker loops are used to provide ticker stream when exchange does not support ws
+//-- tickers loops are used to emulate wsTickers when exchange does not support ws
 
 /**
  * @param {object} changes list of changes to process
@@ -845,7 +871,7 @@ _processSubscriptions(streamClientDescriptor)
  *    "unsubscribe":[{"entity":"","pair":""},...]
  * }
  */
-_processTickerLoop(changes, opt)
+_processTickersLoops(changes, opt)
 {
     // check if we need to unsubscribe
     if (undefined !== changes.unsubscribe)
@@ -854,10 +880,10 @@ _processTickerLoop(changes, opt)
             switch (entry.entity)
             {
                 case 'ticker':
-                    this._unregisterTickerLoop(entry.pair);
+                    this._unregisterTickersLoop(entry.pair);
                     break;
                 case 'tickers':
-                    this._unregisterGlobalTickerLoop();
+                    this._unregisterGlobalTickersLoop();
                     break;
             }
         });
@@ -872,10 +898,10 @@ _processTickerLoop(changes, opt)
                 switch (entry.entity)
                 {
                     case 'ticker':
-                        this._registerTickerLoop(entry.pair);
+                        this._registerTickersLoop(entry.pair);
                         break;
                     case 'tickers':
-                        this._registerGlobalTickerLoop();
+                        this._registerGlobalTickersLoop();
                         break;
                 }
             });
@@ -886,16 +912,16 @@ _processTickerLoop(changes, opt)
 /**
  * Starts a global ticker loop (will be called if exchange supports retrieving tickers for all pairs)
  */
-_registerGlobalTickerLoop()
+_registerGlobalTickersLoop()
 {
     let loop_id = 'global';
     // initialize loop information
-    if (undefined === this._tickerLoop.list[loop_id])
+    if (undefined === this._emulatedWs.wsTickers.list[loop_id])
     {
-        this._tickerLoop.list[loop_id] = {enabled:false,timer:null};
+        this._emulatedWs.wsTickers.list[loop_id] = {enabled:false,timer:null};
     }
     // we already have a loop
-    if (this._tickerLoop.list[loop_id].enabled)
+    if (this._emulatedWs.wsTickers.list[loop_id].enabled)
     {
         return;
     }
@@ -905,22 +931,22 @@ _registerGlobalTickerLoop()
     }
     let self = this;
     let timestamp = Date.now() / 1000.0;
-    this._tickerLoop.list[loop_id].enabled = true;
-    this._tickerLoop.list[loop_id].timestamp = timestamp;
+    this._emulatedWs.wsTickers.list[loop_id].enabled = true;
+    this._emulatedWs.wsTickers.list[loop_id].timestamp = timestamp;
     this._registerConnection(`ticker-${loop_id}`);
-    const getTicker = function(){
+    const doRequest = function(){
         if (debug.enabled)
         {
             debug(`Retrieving tickers for exchange '${self._exchangeId}'`);
         }
         self._exchangeInstance.getTickers().then((data) => {
             // loop has been disabled
-            if (!self._tickerLoop.list[loop_id].enabled)
+            if (!self._emulatedWs.wsTickers.list[loop_id].enabled)
             {
                 return;
             }
             // we already have a newer loop
-            if (self._tickerLoop.list[loop_id].timestamp > timestamp)
+            if (self._emulatedWs.wsTickers.list[loop_id].timestamp > timestamp)
             {
                 return;
             }
@@ -941,34 +967,41 @@ _registerGlobalTickerLoop()
                 }
             });
             // schedule next loop
-            self._tickerLoop.list[loop_id].timer = setTimeout(function(){
-                getTicker();
-            }, self._tickerLoop.period);
+            self._emulatedWs.wsTickers.list[loop_id].timer = setTimeout(function(){
+                doRequest();
+            }, self._emulatedWs.wsTickers.period);
         }).catch((e) => {
             // try again after 5s
             let period = 5000;
-            if (period > self._tickerLoop.period)
+            if (period > self._emulatedWs.wsTickers.period)
             {
-                period = self._tickerLoop.period;
+                period = self._emulatedWs.wsTickers.period;
             }
             logger.warn(`Could not retrieve tickers for exchange '${self._exchangeId}' : will try again in ${period} ms`);
-            logger.warn(JSON.stringify(e));
-            self._tickerLoop.list[loop_id].timer = setTimeout(function(){
-                getTicker();
+            if (e instanceof Errors.BaseError)
+            {
+                logger.warn(JSON.stringify(e));
+            }
+            else
+            {
+                logger.warn(e.stack);
+            }
+            self._emulatedWs.wsTickers.list[loop_id].timer = setTimeout(function(){
+                doRequest();
             }, period);
         });
     }
-    getTicker();
+    doRequest();
 }
 
 /*
  * Stops global ticker loop (will be called if exchange supports retrieving tickers for all pairs)
  */
-_unregisterGlobalTickerLoop()
+_unregisterGlobalTickersLoop()
 {
     let loop_id = 'global';
     // loop is already disabled
-    if (undefined === this._tickerLoop.list[loop_id] || !this._tickerLoop.list[loop_id].enabled)
+    if (undefined === this._emulatedWs.wsTickers.list[loop_id] || !this._emulatedWs.wsTickers.list[loop_id].enabled)
     {
         return;
     }
@@ -977,14 +1010,14 @@ _unregisterGlobalTickerLoop()
         debug(`Stopping global ticker loop for exchange '${this._exchangeId}'`);
     }
     this._unregisterConnection(`ticker-${loop_id}`);
-    this._tickerLoop.list[loop_id].enabled = false;
-    if (null !== this._tickerLoop.list[loop_id].timer)
+    this._emulatedWs.wsTickers.list[loop_id].enabled = false;
+    if (null !== this._emulatedWs.wsTickers.list[loop_id].timer)
     {
-        clearTimeout(this._tickerLoop.list[loop_id].timer);
-        this._tickerLoop.list[loop_id].timer = null;
+        clearTimeout(this._emulatedWs.wsTickers.list[loop_id].timer);
+        this._emulatedWs.wsTickers.list[loop_id].timer = null;
     }
-    this._tickerLoop.list[loop_id].timer = null;
-    this._tickerLoop.list[loop_id].timestamp = Date.now() / 1000.0;
+    this._emulatedWs.wsTickers.list[loop_id].timer = null;
+    this._emulatedWs.wsTickers.list[loop_id].timestamp = Date.now() / 1000.0;
 }
 
 /**
@@ -992,16 +1025,16 @@ _unregisterGlobalTickerLoop()
  *
  * @param {string} pair pair to retrieve ticker for
  */
-_registerTickerLoopForPair(pair)
+_registerTickersLoopForPair(pair)
 {
     let loop_id = pair;
     // initialize loop information
-    if (undefined === this._tickerLoop.list[loop_id])
+    if (undefined === this._emulatedWs.wsTickers.list[loop_id])
     {
-        this._tickerLoop.list[loop_id] = {enabled:false, timer:null};
+        this._emulatedWs.wsTickers.list[loop_id] = {enabled:false, timer:null};
     }
     // we already have a loop
-    if (this._tickerLoop.list[loop_id].enabled)
+    if (this._emulatedWs.wsTickers.list[loop_id].enabled)
     {
         return;
     }
@@ -1011,22 +1044,22 @@ _registerTickerLoopForPair(pair)
     }
     let self = this;
     let timestamp = Date.now() / 1000.0;
-    this._tickerLoop.list[loop_id].enabled = true;
-    this._tickerLoop.list[loop_id].timestamp = timestamp;
+    this._emulatedWs.wsTickers.list[loop_id].enabled = true;
+    this._emulatedWs.wsTickers.list[loop_id].timestamp = timestamp;
     this._registerConnection(`ticker-${loop_id}`);
-    const getTicker = function(){
+    const doRequest = function(){
         if (debug.enabled)
         {
             debug(`Retrieving '${pair}' tickers for exchange '${self._exchangeId}'`);
         }
         self._exchangeInstance.getTickers().then((data) => {
             // loop has been disabled
-            if (!self._tickerLoop.list[loop_id].enabled)
+            if (!self._emulatedWs.wsTickers.list[loop_id].enabled)
             {
                 return;
             }
             // we already have a newer loop
-            if (self._tickerLoop.list[loop_id].timestamp > timestamp)
+            if (self._emulatedWs.wsTickers.list[loop_id].timestamp > timestamp)
             {
                 return;
             }
@@ -1044,36 +1077,43 @@ _registerTickerLoopForPair(pair)
                 self.emit('ticker', evt);
             }
             // schedule next loop
-            self._tickerLoop.list[loop_id].timer = setTimeout(function(){
-                getTicker();
-            }, self._tickerLoop.period);
+            self._emulatedWs.wsTickers.list[loop_id].timer = setTimeout(function(){
+                doRequest();
+            }, self._emulatedWs.wsTickers.period);
         }).catch((e) => {
             // try again after 5s
             let period = 5000;
-            if (period > self._tickerLoop.period)
+            if (period > self._emulatedWs.wsTickers.period)
             {
-                period = self._tickerLoop.period;
+                period = self._emulatedWs.wsTickers.period;
             }
             logger.warn(`Could not retrieve '${pair}' ticker for exchange '${self._exchangeId}' : will try again in ${period} ms`);
-            logger.warn(JSON.stringify(e));
-            self._tickerLoop.list[loop_id].timer = setTimeout(function(){
-                getTicker();
+            if (e instanceof Errors.BaseError)
+            {
+                logger.warn(JSON.stringify(e));
+            }
+            else
+            {
+                logger.warn(e.stack);
+            }
+            self._emulatedWs.wsTickers.list[loop_id].timer = setTimeout(function(){
+                doRequest();
             }, period);
         });
     }
-    getTicker();
+    doRequest();
 }
 
 /**
  * Stops ticker loop for a given pair (will be called if exchange does not support retrieving tickers for all pairs)
  *
- * @param {string} pair pair to retrieve ticker for
+ * @param {string} pair pair to unsubscribe from
  */
-_unregisterTickerLoopForPair(pair)
+_unregisterTickersLoopForPair(pair)
 {
     let loop_id = pair;
     // loop is already disabled
-    if (undefined === this._tickerLoop.list[loop_id] || !this._tickerLoop.list[loop_id].enabled)
+    if (undefined === this._emulatedWs.wsTickers.list[loop_id] || !this._emulatedWs.wsTickers.list[loop_id].enabled)
     {
         return;
     }
@@ -1082,13 +1122,359 @@ _unregisterTickerLoopForPair(pair)
         debug(`Stopping '${pair}' ticker loop for exchange '${this._exchangeId}'`);
     }
     this._unregisterConnection(`ticker-${loop_id}`);
-    this._tickerLoop.list[loop_id].enabled = false;
-    if (null !== this._tickerLoop.list[loop_id].timer)
+    this._emulatedWs.wsTickers.list[loop_id].enabled = false;
+    if (null !== this._emulatedWs.wsTickers.list[loop_id].timer)
     {
-        clearTimeout(this._tickerLoop.list[loop_id].timer);
-        this._tickerLoop.list[loop_id].timer = null;
+        clearTimeout(this._emulatedWs.wsTickers.list[loop_id].timer);
+        this._emulatedWs.wsTickers.list[loop_id].timer = null;
     }
-    this._tickerLoop.list[loop_id].timestamp = Date.now() / 1000.0;
+    this._emulatedWs.wsTickers.list[loop_id].timestamp = Date.now() / 1000.0;
+}
+
+//-- order books loops are used to emulate wsOrderBooks when exchange does not support ws
+
+/**
+ * @param {object} changes list of changes to process
+ * @param {boolean} opt.connect whether or not changes should trigger a connection
+ *
+ *  Each property (subscribe,unsubscribe) is optional
+ *  Entity can be (orderBook)
+ *
+ * {
+ *    "subscribe":[{"entity":"","pair":""},...],
+ *    "unsubscribe":[{"entity":"","pair":""},...]
+ * }
+ */
+_processOrderBooksLoops(changes, opt)
+{
+    // check if we need to unsubscribe
+    if (undefined !== changes.unsubscribe)
+    {
+        _.forEach(changes.unsubscribe, (entry) => {
+            switch (entry.entity)
+            {
+                case 'orderBook':
+                    this._unregisterOrderBookLoop(entry.pair);
+                    break;
+            }
+        });
+    }
+    // check if we need to subscribe
+    if (undefined !== changes.subscribe)
+    {
+        // only if we'be been asked to connect to exchange streams
+        if (opt.connect)
+        {
+            _.forEach(changes.subscribe, (entry) => {
+                switch (entry.entity)
+                {
+                    case 'orderBook':
+                        this._registerOrderBookLoop(entry.pair);
+                        break;
+                }
+            });
+        }
+    }
+}
+
+/**
+ * Starts an order book loop for a given pair
+ *
+ * @param {string} pair pair to retrieve order book for
+ */
+_registerOrderBookLoop(pair)
+{
+    let loop_id = pair;
+    // initialize loop information
+    if (undefined === this._emulatedWs.wsOrderBooks.list[loop_id])
+    {
+        this._emulatedWs.wsOrderBooks.list[loop_id] = {enabled:false, timer:null};
+    }
+    // we already have a loop
+    if (this._emulatedWs.wsOrderBooks.list[loop_id].enabled)
+    {
+        return;
+    }
+    if (debug.enabled)
+    {
+        debug(`Starting '${pair}' order book loop for exchange '${this._exchangeId}'`);
+    }
+    let self = this;
+    let timestamp = Date.now() / 1000.0;
+    this._emulatedWs.wsOrderBooks.list[loop_id].enabled = true;
+    this._emulatedWs.wsOrderBooks.list[loop_id].timestamp = timestamp;
+    this._registerConnection(`orderBook-${loop_id}`);
+    const doRequest = function(){
+        if (debug.enabled)
+        {
+            debug(`Retrieving '${pair}' order book for exchange '${self._exchangeId}'`);
+        }
+        self._exchangeInstance.getOrderBook(pair).then((data) => {
+            // loop has been disabled
+            if (!self._emulatedWs.wsOrderBooks.list[loop_id].enabled)
+            {
+                return;
+            }
+            // we already have a newer loop
+            if (self._emulatedWs.wsOrderBooks.list[loop_id].timestamp > timestamp)
+            {
+                return;
+            }
+            if (debug.enabled)
+            {
+                debug(`Got new '${pair}' order book for exchange '${self._exchangeId}'`);
+            }
+            let evt = {
+                exchange:self._exchangeId,
+                pair:pair,
+                cseq:Date.now(),
+                data:{
+                    buy:data.buy,
+                    sell:data.sell
+                }
+            }
+            self.emit('orderBook', evt);
+            // schedule next loop
+            self._emulatedWs.wsOrderBooks.list[loop_id].timer = setTimeout(function(){
+                doRequest();
+            }, self._emulatedWs.wsOrderBooks.period);
+        }).catch((e) => {
+            // try again after 5s
+            let period = 5000;
+            if (period > self._emulatedWs.wsOrderBooks.period)
+            {
+                period = self._emulatedWs.wsOrderBooks.period;
+            }
+            logger.warn(`Could not retrieve '${pair}' order book for exchange '${self._exchangeId}' : will try again in ${period} ms`);
+            if (e instanceof Errors.BaseError)
+            {
+                logger.warn(JSON.stringify(e));
+            }
+            else
+            {
+                logger.warn(e.stack);
+            }
+            self._emulatedWs.wsOrderBooks.list[loop_id].timer = setTimeout(function(){
+                doRequest();
+            }, period);
+        });
+    }
+    doRequest();
+}
+
+/**
+ * Stops order book loop for a given pair
+ *
+ * @param {string} pair pair to unsubscribe from
+ */
+_unregisterOrderBookLoop(pair)
+{
+    let loop_id = pair;
+    // loop is already disabled
+    if (undefined === this._emulatedWs.wsOrderBooks.list[loop_id] || !this._emulatedWs.wsOrderBooks.list[loop_id].enabled)
+    {
+        return;
+    }
+    if (debug.enabled)
+    {
+        debug(`Stopping '${pair}' order book loop for exchange '${this._exchangeId}'`);
+    }
+    this._unregisterConnection(`orderBook-${loop_id}`);
+    this._emulatedWs.wsOrderBooks.list[loop_id].enabled = false;
+    if (null !== this._emulatedWs.wsOrderBooks.list[loop_id].timer)
+    {
+        clearTimeout(this._emulatedWs.wsOrderBooks.list[loop_id].timer);
+        this._emulatedWs.wsOrderBooks.list[loop_id].timer = null;
+    }
+    this._emulatedWs.wsOrderBooks.list[loop_id].timestamp = Date.now() / 1000.0;
+}
+
+//-- trades loops are used to emulate wsTrades when exchange does not support ws
+
+/**
+ * @param {object} changes list of changes to process
+ * @param {boolean} opt.connect whether or not changes should trigger a connection
+ *
+ *  Each property (subscribe,unsubscribe) is optional
+ *  Entity can be (trades)
+ *
+ * {
+ *    "subscribe":[{"entity":"","pair":""},...],
+ *    "unsubscribe":[{"entity":"","pair":""},...]
+ * }
+ */
+_processTradesLoops(changes, opt)
+{
+    // check if we need to unsubscribe
+    if (undefined !== changes.unsubscribe)
+    {
+        _.forEach(changes.unsubscribe, (entry) => {
+            switch (entry.entity)
+            {
+                case 'trades':
+                    this._unregisterTradesLoop(entry.pair);
+                    break;
+            }
+        });
+    }
+    // check if we need to subscribe
+    if (undefined !== changes.subscribe)
+    {
+        // only if we'be been asked to connect to exchange streams
+        if (opt.connect)
+        {
+            _.forEach(changes.subscribe, (entry) => {
+                switch (entry.entity)
+                {
+                    case 'trades':
+                        this._registerTradesLoop(entry.pair);
+                        break;
+                }
+            });
+        }
+    }
+}
+
+/**
+ * Starts a trades loop for a given pair
+ *
+ * @param {string} pair pair to retrieve trades for
+ */
+_registerTradesLoop(pair)
+{
+    let loop_id = pair;
+    // initialize loop information
+    if (undefined === this._emulatedWs.wsTrades.list[loop_id])
+    {
+        this._emulatedWs.wsTrades.list[loop_id] = {enabled:false, timer:null, lastTrade:null};
+    }
+    // we already have a loop
+    if (this._emulatedWs.wsTrades.list[loop_id].enabled)
+    {
+        return;
+    }
+    if (debug.enabled)
+    {
+        debug(`Starting '${pair}' trades loop for exchange '${this._exchangeId}'`);
+    }
+    let self = this;
+    let timestamp = Date.now() / 1000.0;
+    this._emulatedWs.wsTrades.list[loop_id].enabled = true;
+    this._emulatedWs.wsTrades.list[loop_id].timestamp = timestamp;
+    this._emulatedWs.wsTrades.list[loop_id].lastTrade = null;
+    this._registerConnection(`orderBook-${loop_id}`);
+    const doRequest = function(){
+        if (debug.enabled)
+        {
+            debug(`Retrieving '${pair}' trades for exchange '${self._exchangeId}'`);
+        }
+        self._exchangeInstance.getTrades(pair).then((data) => {
+            // loop has been disabled
+            if (!self._emulatedWs.wsTrades.list[loop_id].enabled)
+            {
+                return;
+            }
+            // we already have a newer loop
+            if (self._emulatedWs.wsTrades.list[loop_id].timestamp > timestamp)
+            {
+                return;
+            }
+            if (debug.enabled)
+            {
+                debug(`Got new '${pair}' trades for exchange '${self._exchangeId}'`);
+            }
+            let evt = {
+                exchange:self._exchangeId,
+                pair:pair,
+                data:[]
+            };
+            _.forEach(data, (trade) => {
+                if (null === self._emulatedWs.wsTrades.list[loop_id].lastTrade)
+                {
+                    evt.data.push(trade);
+                    return;
+                }
+                // only keep trade if id is greater than previous
+                if (null !== self._emulatedWs.wsTrades.list[loop_id].lastTrade.id && null !== trade.id)
+                {
+                    // ignore trade
+                    if (trade.id <= self._emulatedWs.wsTrades.list[loop_id].lastTrade.id)
+                    {
+                        return;
+                    }
+                }
+                // compare timestamps
+                else
+                {
+                    if (trade.timestamp <= self._emulatedWs.wsTrades.list[loop_id].lastTrade.timestamp)
+                    {
+                        return;
+                    }
+                }
+                evt.data.push(trade);
+            });
+            // update last trade with newest one
+            if (0 != evt.data.length)
+            {
+                self._emulatedWs.wsTrades.list[loop_id].lastTrade = {
+                    id:evt.data[0].id,
+                    timestamp:evt.data[0].timestamp
+                }
+            }
+            self.emit('trades', evt);
+            // schedule next loop
+            self._emulatedWs.wsTrades.list[loop_id].timer = setTimeout(function(){
+                doRequest();
+            }, self._emulatedWs.wsTrades.period);
+        }).catch((e) => {
+            // try again after 5s
+            let period = 5000;
+            if (period > self._emulatedWs.wsTrades.period)
+            {
+                period = self._emulatedWs.wsTrades.period;
+            }
+            logger.warn(`Could not retrieve '${pair}' trades for exchange '${self._exchangeId}' : will try again in ${period} ms`);
+            if (e instanceof Errors.BaseError)
+            {
+                logger.warn(JSON.stringify(e));
+            }
+            else
+            {
+                logger.warn(e.stack);
+            }
+            self._emulatedWs.wsTrades.list[loop_id].timer = setTimeout(function(){
+                doRequest();
+            }, period);
+        });
+    }
+    doRequest();
+}
+
+/**
+ * Stops trades loop for a given pair
+ *
+ * @param {string} pair pair to unsubscribe from
+ */
+_unregisterTradesLoop(pair)
+{
+    let loop_id = pair;
+    // loop is already disabled
+    if (undefined === this._emulatedWs.wsTrades.list[loop_id] || !this._emulatedWs.wsTrades.list[loop_id].enabled)
+    {
+        return;
+    }
+    if (debug.enabled)
+    {
+        debug(`Stopping '${pair}' trades loop for exchange '${this._exchangeId}'`);
+    }
+    this._unregisterConnection(`trades-${loop_id}`);
+    this._emulatedWs.wsTrades.list[loop_id].enabled = false;
+    if (null !== this._emulatedWs.wsTrades.list[loop_id].timer)
+    {
+        clearTimeout(this._emulatedWs.wsTrades.list[loop_id].timer);
+        this._emulatedWs.wsTrades.list[loop_id].timer = null;
+    }
+    this._emulatedWs.wsTrades.list[loop_id].timestamp = Date.now() / 1000.0;
 }
 
 }