+++ app/exchanges/kucoin/exchange.js
@@ -31,9 +31,9 @@ const defaultKlinesInterval = '5m';
 // list of all possible features (should be enabled by default if supported by class)
 const supportedFeatures = {
     'pairs':{enabled:true},
-    'tickers':{enabled:true, withoutPair:true}, 'wsTickers':{enabled:true,'emulated':true},
-    'orderBooks':{enabled:true}, 'wsOrderBooks':{enabled:false},
-    'trades':{enabled:true}, 'wsTrades':{enabled:false},
+    'tickers':{enabled:true, withoutPair:true}, 'wsTickers':{enabled:true,emulated:true},
+    'orderBooks':{enabled:true}, 'wsOrderBooks':{enabled:true,emulated:true},
+    'trades':{enabled:true}, 'wsTrades':{enabled:true,emulated:true},
     'klines':{enabled:true,intervals:supportedKlinesIntervals,defaultInterval:defaultKlinesInterval}, 'wsKlines':{enabled:false},
     'orders':{enabled:true, withoutPair:false},
     'openOrders':{enabled:true, withoutPair:false},